backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.4
scipy>=1.12.0
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
pydantic>=2.10.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
